Added an abstraction level to RdsResultStorage.
